Unified the search to use keyword analyzers instead of whitespace analyzers. Nonexact fields will still use Voikko analyzer that tokenizes on white space but exact fields should all now be indexed and searched using keyword analyzer.

+++ metka/src/main/java/fi/uta/fsd/metka/mvc/services/impl/ExpertSearchServiceImpl.java
@@ -107,7 +107,7 @@ public class ExpertSearchServiceImpl implements ExpertSearchService {
                 Logger.warning(getClass(), "Revisionable was not found for id "+result.getId());
                 continue;
             }
-            Pair<ReturnResult, RevisionData> pair = revisions.getRevisionData(result.getId(), result.getNo().intValue());
+            Pair<ReturnResult, RevisionData> pair = revisions.getRevisionData(result.getId(), result.getNo());
             if(pair.getLeft() != ReturnResult.REVISION_FOUND) {
                 Logger.warning(getClass(), "Couldn't find a revision for search result "+result.toString());
                 continue;
